# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Primeiro Prontuário API
  version: 1.0.0
  description: |
    Contrato completo do MVP acadêmico para um único médico.
    Use somente dados fictícios ou anonimizados.
  license:
    name: Não licenciado - uso acadêmico demonstrativo
    identifier: LicenseRef-Proprietary
servers:
  - url: https://prontuario.local
    description: Cluster Kubernetes local com TLS confiado apenas na máquina de demonstração
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/patients/{patientId}/attachments:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    get:
      tags: [Anexos]
      operationId: listPatientAttachments
      summary: Listar metadados dos anexos do paciente
      responses:
        "200":
          description: Anexos ativos e lápides de removidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    post:
      tags: [Anexos]
      operationId: uploadAttachment
      summary: Enviar anexo PDF, JPG, PNG ou Markdown de até 10 MiB
      security:
        - sessionCookie: []
          csrfHeader: []
      parameters:
        - name: consultationId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Anexo persistido
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "413":
          $ref: "#/components/responses/Problem"
        "415":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/attachments/{attachmentId}:
    parameters:
      - $ref: "#/components/parameters/AttachmentId"
    get:
      tags: [Anexos]
      operationId: getAttachmentMetadata
      summary: Consultar metadados ou lápide do anexo
      responses:
        "200":
          description: Metadados do anexo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      tags: [Anexos]
      operationId: removeAttachment
      summary: Remover binário e manter lápide
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveAttachmentRequest"
      responses:
        "204":
          description: Anexo removido ou remoção idempotente concluída
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/attachments/{attachmentId}/content:
    parameters:
      - $ref: "#/components/parameters/AttachmentId"
    get:
      tags: [Anexos]
      operationId: downloadAttachment
      summary: Baixar binário autenticado
      responses:
        "200":
          description: Conteúdo com Content-Disposition attachment
          headers:
            Content-Disposition:
              schema:
                type: string
            Content-Length:
              schema:
                type: integer
                format: int64
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            text/markdown:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "410":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessão server-side; HttpOnly, SameSite=Lax e Secure em HTTPS
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: |
        Token devolvido por GET /api/v1/auth/csrf e associado ao cookie
        XSRF-TOKEN do mesmo cliente. Exigido em mutações autenticadas.
  headers:
    CorrelationId:
      description: UUID de correlação aceito ou gerado pela API
      schema:
        type: string
        format: uuid
  parameters:
...
    PatientId:
      name: patientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    AttachmentId:
      name: attachmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: Problem Details seguro; status e tipo dependem da falha
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    AttachmentStatus:
      type: string
      enum: [ACTIVE, REMOVED]
    Attachment:
      type: object
      required: [id, patientId, mediaType, size, sha256, status]
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        consultationId:
          type: [string, "null"]
          format: uuid
        originalFilename:
          type: [string, "null"]
          description: Omitido na lápide removida
        mediaType:
          type: string
          enum: [application/pdf, image/jpeg, image/png, text/markdown]
        size:
          type: integer
          format: int64
          minimum: 1
          maximum: 10485760
        sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        status:
          $ref: "#/components/schemas/AttachmentStatus"
        uploadedBy:
          type: [string, "null"]
          format: uuid
        createdAt:
          type: [string, "null"]
          format: date-time
        removedBy:
          type: [string, "null"]
          format: uuid
        removalJustification:
          type: [string, "null"]
        removedAt:
          type: [string, "null"]
          format: date-time
    RemoveAttachmentRequest:
      type: object
      required: [justification]
      additionalProperties: false
      properties:
        justification:
          type: string
          minLength: 1
...
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    Problem:
      type: object
      required: [type, title, status, detail, correlationId]
      properties:
        type:
          type: string
          format: uri
          examples: [urn:problem:invalid-request]
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        correlationId:
          type: string
          format: uuid
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"